requirement.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
certifi==2021.5.30
charset-normalizer==2.0.6; python_version >= '3'
click==8.0.1; python_version >= '3.6'
cython==0.29.24; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
flask-sqlalchemy==2.5.1
flask==2.0.1
greenlet==1.1.1; python_version >= '3' and platform_machine in 'x86_64 X86_64 aarch64 AARCH64 ppc64le PPC64LE amd64 AMD64 win32 WIN32'
idna==3.2; python_version >= '3'
itsdangerous==2.0.1; python_version >= '3.6'
jinja2==3.0.1; python_version >= '3.6'
markupsafe==2.0.1; python_version >= '3.6'
mysql-connector-python==8.0.26
numpy==1.21.2; python_version < '3.11' and python_version >= '3.7'
pkuseg==0.0.25
protobuf==3.18.0
requests==2.26.0
sqlalchemy==1.4.23; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
urllib3==1.26.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
werkzeug==2.0.1; python_version >= '3.6'
